PumpOS/backend/requirements.txt
aioquic==1.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==25.3.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
cryptography==44.0.3
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
facenet-pytorch==2.5.3
Faker==37.4.0
filelock==3.13.1
Flask==3.1.0
fsspec==2024.6.1
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
gunicorn==23.0.0
h11==0.16.0
h2==4.1.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kaitaistruct==0.10
ldap3==2.9.1
MarkupSafe==3.0.2
mitmproxy==12.1.1
mitmproxy-windows==0.12.6
mitmproxy_rs==0.12.6
mpmath==1.3.0
msgpack==1.1.0
networkx==3.3
numpy==2.1.2
opencv-python==4.12.0.88
packaging==24.2
passlib==1.7.4
pillow==11.1.0
promise==2.3
psycopg2-binary==2.9.10
publicsuffix2==2.20191221
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydivert==2.1.0
PyJWT==2.9.0
pylsqpack==0.3.22
pyOpenSSL==25.0.0
pyparsing==3.2.3
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
qrcode==8.1
razorpay==1.4.2
requests==2.32.3
ruamel.yaml==0.18.10
service-identity==24.2.0
setuptools==78.1.0
six==1.17.0
sortedcontainers==2.4.0
sqlparse==0.5.3
sympy==1.13.3
text-unidecode==1.3
torch==2.8.0
torchvision==0.23.0
tornado==6.5
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.3.0
ultralytics==8.3.203
ultralytics-thop==2.0.17
urwid==2.6.16
wcwidth==0.2.13
Werkzeug==3.1.3
wsproto==1.2.0
zstandard==0.23.0
